fix: resolve static analysis findings
